Pretty window titles in TextMate 2
http://blog.macromates.com/2011/format-strings/

I sure wish I could figure out how to get `windowTitle` to accept a variable. Replacing `trey` with `$USER` doesn't do the trick for some reason.

From the example on the blog post, I added a dash `[\w-]` so that it would match a directory with a dash in it.