Let’s see if we can trust oh-my-zsh now.
Please don’t mess this up again.

I would need to rename this to be able to continue to use it with the
git plugin installed. I like the little echo of what it’s going to do,
but I think I can handle not seeing it now. I’m glad my `ggpushf`
isn’t something that’s in the git plugin aliases.